refactor(entities-plugins): datakit rename and adjust file structure
