+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/FlowPanels.vue
@@ -81,8 +81,8 @@ import { useDraggable, useElementBounding } from '@vueuse/core'
 import { useVueFlow } from '@vue-flow/core'
 import { computed, nextTick, ref, useId, useTemplateRef, watch } from 'vue'
 
-import { useEditorStore } from '../../composables'
-import FlowCanvas from './EditorCanvasFlow.vue'
+import { useEditorStore } from '../composables'
+import FlowCanvas from './FlowCanvas.vue'
 
 const { readonly, resizable } = defineProps<{
   readonly?: boolean
